Explode classes into files and repackage
