+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/model/InternalModel.kt
@@ -1,0 +1,8 @@
+package net.twisterrob.ghlint.model
+
+import org.snakeyaml.engine.v2.nodes.MappingNode
+
+internal interface InternalModel : Model {
+
+	val node: MappingNode
+}
