+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/model/Job.kt
@@ -1,0 +1,56 @@
+package net.twisterrob.ghlint.model
+
+import net.twisterrob.ghlint.yaml.array
+import net.twisterrob.ghlint.yaml.getOptional
+import net.twisterrob.ghlint.yaml.getOptionalText
+import net.twisterrob.ghlint.yaml.getRequired
+import net.twisterrob.ghlint.yaml.getRequiredText
+import org.snakeyaml.engine.v2.nodes.MappingNode
+
+public sealed class Job protected constructor(
+) : InternalModel {
+
+	public abstract val parent: Workflow
+	public abstract val id: String
+	public val name: String?
+		get() = node.getOptionalText("name")
+
+	public class NormalJob internal constructor(
+		public override val parent: Workflow,
+		override val id: String,
+		override val node: MappingNode,
+	) : Job() {
+
+		public val steps: List<Step>
+			get() = node.getRequired("steps").array.map { Step.from(this, it as MappingNode) }
+
+		public val defaults: Defaults?
+			get() = node.getOptional("defaults")?.let { Defaults.from(it as MappingNode) }
+
+		public class Defaults internal constructor(
+			private val node: MappingNode,
+		) {
+
+			public val shell: String?
+				get() = node.getOptionalText("shell")
+
+			public companion object
+		}
+
+		public companion object
+	}
+
+	public class ReusableWorkflowCallJob internal constructor(
+		public override val parent: Workflow,
+		override val id: String,
+		override val node: MappingNode,
+	) : Job() {
+
+		public val uses: String
+			get() = node.getRequiredText("uses")
+
+		public companion object
+	}
+
+	public companion object
+}
