+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/model/Model.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.ghlint.model
+
+public sealed interface Model {
+	public companion object
+}
