--- modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/model/Rule.kt
@@ -1,6 +1,0 @@
-package net.twisterrob.ghlint.model
-
-public interface Rule {
-
-	public fun check(workflow: Workflow): List<Finding>
-}
