+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/model/Workflow.kt
@@ -1,45 +1,10 @@
 package net.twisterrob.ghlint.model
 
-import net.twisterrob.ghlint.yaml.array
-import net.twisterrob.ghlint.yaml.getOptional
 import net.twisterrob.ghlint.yaml.getOptionalText
 import net.twisterrob.ghlint.yaml.getRequired
-import net.twisterrob.ghlint.yaml.getRequiredText
 import net.twisterrob.ghlint.yaml.map
 import net.twisterrob.ghlint.yaml.text
 import org.snakeyaml.engine.v2.nodes.MappingNode
-import java.nio.file.Path
-import kotlin.io.path.name
-
-public class File internal constructor(
-	public val file: FileName,
-) {
-
-	internal fun readText(): String =
-		java.io.File(file.path).readText()
-
-	public companion object
-}
-
-@JvmInline
-public value class FileName(
-	public val path: String,
-) {
-
-	public val name: String
-		get() = Path.of(path).name
-
-	public companion object
-}
-
-public sealed interface Model {
-	public companion object
-}
-
-internal interface InternalModel : Model {
-
-	val node: MappingNode
-}
 
 public class Workflow internal constructor(
 	public val parent: File,
@@ -53,99 +18,6 @@ public class Workflow internal constructor(
 		get() = node.getRequired("jobs").map
 			.mapKeys { (key, _) -> key.text }
 			.mapValues { (key, value) -> Job.from(this, key, value as MappingNode) }
-
-	public companion object
-}
-
-public sealed class Job protected constructor(
-) : InternalModel {
-
-	public abstract val parent: Workflow
-	public abstract val id: String
-	public val name: String?
-		get() = node.getOptionalText("name")
-
-	public class NormalJob internal constructor(
-		public override val parent: Workflow,
-		override val id: String,
-		override val node: MappingNode,
-	) : Job() {
-
-		public val steps: List<Step>
-			get() = node.getRequired("steps").array.map { Step.from(this, it as MappingNode) }
-
-		public val defaults: Defaults?
-			get() = node.getOptional("defaults")?.let { Defaults.from(it as MappingNode) }
-
-		public class Defaults internal constructor(
-			private val node: MappingNode,
-		) {
-
-			public val shell: String?
-				get() = node.getOptionalText("shell")
-
-			public companion object
-		}
-
-		public companion object
-	}
-
-	public class ReusableWorkflowCallJob internal constructor(
-		public override val parent: Workflow,
-		override val id: String,
-		override val node: MappingNode,
-	) : Job() {
-
-		public val uses: String
-			get() = node.getRequiredText("uses")
-
-		public companion object
-	}
-
-	public companion object
-}
-
-public sealed class Step protected constructor(
-) : InternalModel {
-
-	public abstract val parent: Job.NormalJob
-
-	public val name: String?
-		get() = node.getOptionalText("name")
-
-	public val id: String?
-		get() = node.getOptionalText("id")
-
-	@Suppress("detekt.VariableNaming")
-	public val `if`: String?
-		get() = node.getOptionalText("if")
-
-	public class Run internal constructor(
-		public override val parent: Job.NormalJob,
-		override val node: MappingNode,
-	) : Step() {
-
-		@Suppress("detekt.MemberNameEqualsClassName")
-		public val run: String
-			get() = node.getRequiredText("run")
-
-		public val shell: String?
-			get() = node.getOptionalText("shell")
-
-		public companion object
-	}
-
-	public class Uses internal constructor(
-		public override val parent: Job.NormalJob,
-		override val node: MappingNode,
-	) : Step() {
-
-		@Suppress("detekt.MemberNameEqualsClassName")
-		public val uses: String
-			get() = node.getRequiredText("uses")
-
-		public companion object
-	}
 
 	public companion object
 }
