+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/results/ColumnNumber.kt
@@ -1,0 +1,9 @@
+package net.twisterrob.ghlint.results
+
+@JvmInline
+public value class ColumnNumber(
+	public val number: Int,
+) {
+
+	public companion object
+}
