+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/results/Position.kt
@@ -1,0 +1,9 @@
+package net.twisterrob.ghlint.results
+
+public class Position(
+	public val line: LineNumber,
+	public val column: ColumnNumber,
+) {
+
+	public companion object
+}
