+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/rule/Reporting.kt
@@ -1,0 +1,8 @@
+package net.twisterrob.ghlint.rule
+
+import net.twisterrob.ghlint.model.Model
+
+public interface Reporting {
+
+	public fun report(issue: Issue, context: Model)
+}
