+++ modules/ghlint/src/main/kotlin/net/twisterrob/ghlint/rules/MandatoryShellRule.kt
@@ -1,9 +1,9 @@
 package net.twisterrob.ghlint.rules
 
-import net.twisterrob.ghlint.model.Issue
-import net.twisterrob.ghlint.model.Reporting
 import net.twisterrob.ghlint.model.Step
-import net.twisterrob.ghlint.model.VisitorRule
+import net.twisterrob.ghlint.rule.Issue
+import net.twisterrob.ghlint.rule.Reporting
+import net.twisterrob.ghlint.rule.VisitorRule
 
 public class MandatoryShellRule : VisitorRule {
